refactor(history): extract reusable edit screen
